change(cl): add app_name to templates in store, small fixes, relocated favourites to accounts
