+++ accounts/templates/hello.html
@@ -1,0 +1,39 @@
+{% extends 'main.html' %}
+{% load static %}
+{% block content %}
+
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>HollyMovies</title>
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/css/bootstrap.min.css" integrity="sha384-GJzZqFGwb1QTTN6wy59ffF1BuGJpLSa9DkKMp0DgiMDm4iYMj70gZWKYbI706tWS" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.6/umd/popper.min.js" integrity="sha384-wHAiFfRlMFy6i5SRaxvfOCifBUQy1xHdJ/yoi7FRNXMRBu5WHdZYu1hA6ZOblgut" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.2.1/js/bootstrap.min.js" integrity="sha384-B0UglyR+jN6CkvvICOB2joaf5I4l3gm9GU6Hc1og6Ls7i6U/mkkaduKaBhlAXv9k" crossorigin="anonymous"></script>
+</head>
+
+<div>
+    <h1>Hello!</h1>
+    <div class="nav-item dropdown">
+        {% if user.is_authenticated %}
+          <a class="nav-link active dropdown-toggle" href="#" data-toggle="dropdown">
+            {% if user.first_name %}Hello, {{ user.first_name }}!
+            {% else %}Hello, {{ user.username }}!{% endif %}
+          </a>
+          <div class="dropdown-menu dropdown-menu-right">
+            <a class="dropdown-item text-right" href="{% url 'accounts:logout' %}">Logout</a>
+            <a class="dropdown-item text-right" href="{% url 'accounts:password_change' %}">Change password</a>
+          </div>
+        {% else %}
+          <a class="nav-link active dropdown-toggle" href="#" data-toggle="dropdown">
+            You are not logged in.
+          </a>
+          <div class="dropdown-menu dropdown-menu-right">
+            <a class="dropdown-item text-right" href="{% url 'accounts:login' %}">Login</a>
+            <a class="dropdown-item text-right" href="{% url 'accounts:sign_up' %}">Sign up</a>
+          </div>
+        {% endif %}
+    </div>
+</div>
+
+{% endblock content %}
