+++ accounts/templates/order_history.html
@@ -4,7 +4,7 @@
 
 <div>
     <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
-    <p><a href="{% url 'store:favourites' %}">Favourites</a></p>
+    <p><a href="{% url 'accounts:favourites' %}">Favourites</a></p>
     <p>SIMPLE NAV JUST FOR NOW</p>
     <p>Products in Cart: {{ cart_items }}</p>
