+++ accounts/templates/orders_history.html
@@ -6,7 +6,7 @@
     <p><a href="{% url 'accounts:profile' %}">Profile</a></p>
     <p>SIMPLE NAV JUST FOR NOW</p>
     <p><a href="{% url 'accounts:orders_history' %}">Orders History</a></p>
-    <p><a href="{% url 'store:favourites' %}">Favourites</a></p>
+    <p><a href="{% url 'accounts:favourites' %}">Favourites</a></p>
     <p>Products in Cart: {{ cart_items }}</p>
 
     <div>
