+++ store/templates/categories.html
@@ -8,7 +8,7 @@
             {% for category in object_list %}
                 <div>
                     <img src="{{ category.imageURL }}">
-                    <p><strong><a href="{% url 'category' category.id %}">{{ category.name }}</a></strong></p>
+                    <p><strong><a href="{% url 'store:category' category.id %}">{{ category.name }}</a></strong></p>
                 </div>
             {% endfor %}
         </div>
