+++ store/templates/main.html
@@ -98,7 +98,7 @@
                         <a href="{% url 'store:cart' %}">Koszyk</a>
                       </li>
                       <li>
-                        <a href="{% url 'accounts:login' %}">Moje konto</a>
+                        <a href="{% url 'accounts:hello' %}">Moje konto</a>
                       </li>
                       <li>
                         <a href="{% url 'store:checkout' %}">Podsumowanie Zakupów</a>
