+++ store/urls.py
@@ -15,7 +15,6 @@ urlpatterns = [
     path('update_item/', update_item, name="update_item"),
     path('process_order/', process_order, name="process_order"),
     path('<int:meta_product_id>/meta_product/', meta_product, name="meta_product"),
-    path('favourites/', favourites, name='favourites'),
   
 ]
